ref(build): Reenable lambda layer release in craft (#5207)
