+++ packages/serverless/src/awslambda-auto.ts
@@ -6,7 +6,11 @@ if (lambdaTaskRoot) {
   if (!handlerString) {
     throw Error(`LAMBDA_TASK_ROOT is non-empty(${lambdaTaskRoot}) but _HANDLER is not set`);
   }
-  Sentry.AWSLambda.init();
+
+  Sentry.AWSLambda.init({
+    _invokedByLambdaLayer: true,
+  });
+
   Sentry.AWSLambda.tryPatchHandler(lambdaTaskRoot, handlerString);
 } else {
   throw Error('LAMBDA_TASK_ROOT environment variable is not set');
