+++ packages/serverless/tsconfig.json
@@ -5,6 +5,6 @@
 
   "compilerOptions": {
     // package-specific options
-    "target": "ES2018",
+    "target": "ES2018"
   }
 }
